Create destination folders for certificates if not yet existing

+++ Certificates/clear_certificates.py
@@ -12,12 +12,13 @@ def clear_certificates():
     ]
 
     for path in folder:
-        for file in os.listdir(path):
-            file_endings = [".key", ".txt", ".crt", ".pem", ".crl"]  # Add the desired file endings here
-            
-            if any(file.endswith(ending) for ending in file_endings):
-                print(f"Removing {file}")
-                os.remove(os.path.join(path, file))
+        if(os.path.exists(path)):
+            for file in os.listdir(path):
+                file_endings = [".key", ".txt", ".crt", ".pem", ".crl"]  # Add the desired file endings here
+                
+                if any(file.endswith(ending) for ending in file_endings):
+                    print(f"Removing {file}")
+                    os.remove(os.path.join(path, file))
 
 if __name__ == "__main__":
     clear_certificates()
